# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: workout
    description: AI workout generation and saved plans
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /workout/plans:
    get:
      operationId: listWorkoutPlans
      tags: [workout]
      summary: List all saved workout plans
      responses:
        "200":
          description: List of saved workout plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkoutPlan"
    post:
      operationId: saveWorkoutPlan
      tags: [workout]
      summary: Save a generated workout plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkoutPlanInput"
      responses:
        "201":
          description: Saved plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutPlan"

  /workout/plans/{id}:
    get:
      operationId: getWorkoutPlan
      tags: [workout]
      summary: Get a single saved plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: The workout plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutPlan"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteWorkoutPlan
      tags: [workout]
      summary: Delete a saved plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /workout/generate:
    post:
      operationId: generateWorkout
      tags: [workout]
      summary: Generate an AI workout plan (SSE streaming)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkoutRequest"
      responses:
        "200":
          description: SSE stream of workout content chunks
          content:
            text/event-stream: {}

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    StrengthAssessment:
      type: object
      description: Results from the athlete's strength assessment tests
      properties:
        squat5RM:
          type: string
          description: "e.g. '135 lbs x 5' or '100 kg x 3'"
        deadlift5RM:
          type: string
          description: "e.g. '225 lbs x 5'"
        bench5RM:
          type: string
          description: "e.g. '115 lbs x 5'"
        overheadPress5RM:
          type: string
          description: "e.g. '75 lbs x 5'"
        maxPushUps:
          type: integer
          description: Maximum consecutive push-ups
        bodyweightSquats:
          type: integer
          description: Bodyweight squats completed in 1 minute
        verticalJump:
          type: string
          description: "e.g. '24 inches' (optional)"
        sprintTime:
          type: string
          description: "e.g. '4.8s for 40 yards' (optional)"

    WorkoutPlan:
      type: object
      properties:
        id:
          type: integer
        sport:
          type: string
        experienceLevel:
          type: string
        age:
          type: integer
        weight:
          type: ["string", "null"]
        currentStats:
          type: string
        goals:
          type: string
        daysPerWeek:
          type: integer
        planContent:
          type: string
        strengthAssessment:
          $ref: "#/components/schemas/StrengthAssessment"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - sport
        - experienceLevel
        - age
        - currentStats
        - goals
        - daysPerWeek
        - planContent
        - createdAt

    WorkoutPlanInput:
      type: object
      properties:
        sport:
          type: string
        experienceLevel:
          type: string
        age:
          type: integer
        weight:
          type: string
        currentStats:
          type: string
        goals:
          type: string
        daysPerWeek:
          type: integer
        planContent:
          type: string
        strengthAssessment:
          $ref: "#/components/schemas/StrengthAssessment"
      required:
        - sport
        - experienceLevel
        - age
        - currentStats
        - goals
        - daysPerWeek
        - planContent

    WorkoutRequest:
      type: object
      properties:
        sport:
          type: string
        experienceLevel:
          type: string
        age:
          type: integer
        weight:
          type: string
        currentStats:
          type: string
        goals:
          type: string
        daysPerWeek:
          type: integer
        strengthAssessment:
          $ref: "#/components/schemas/StrengthAssessment"
      required:
        - sport
        - experienceLevel
        - age
        - currentStats
        - goals
        - daysPerWeek

    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error