# lib/api-zod/src/generated/types/strengthAssessment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * Results from the athlete's strength assessment tests
 */
export interface StrengthAssessment {
  /** e.g. '135 lbs x 5' or '100 kg x 3' */
  squat5RM?: string;
  /** e.g. '225 lbs x 5' */
  deadlift5RM?: string;
  /** e.g. '115 lbs x 5' */
  bench5RM?: string;
  /** e.g. '75 lbs x 5' */
  overheadPress5RM?: string;
  /** Maximum consecutive push-ups */
  maxPushUps?: number;
  /** Bodyweight squats completed in 1 minute */
  bodyweightSquats?: number;
  /** e.g. '24 inches' (optional) */
  verticalJump?: string;
  /** e.g. '4.8s for 40 yards' (optional) */
  sprintTime?: string;
}