# lib/api-zod/src/generated/types/workoutPlanInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StrengthAssessment } from './strengthAssessment';

export interface WorkoutPlanInput {
  sport: string;
  experienceLevel: string;
  age: number;
  weight?: string;
  currentStats: string;
  goals: string;
  daysPerWeek: number;
  planContent: string;
  strengthAssessment?: StrengthAssessment;
}